backend/requirements.txt
django>=4.2.0
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
python-decouple
drf-nested-routers